# be/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error (e.g. email already taken)"
          },
          "429": {
            "description": "Too many requests"
          }
        },
        "summary": "Register a new user",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          },
          "429": {
            "description": "Too many requests"
          }
        },
        "summary": "Login with email and password",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/api/v1/auth/google/exchange": {
      "post": {
        "operationId": "AuthController_googleExchange",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleExchangeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tokens returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired code"
          }
        },
        "summary": "Exchange one-time OAuth code for tokens",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tokens refreshed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired refresh token"
          }
        },
        "summary": "Refresh access token using a refresh token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logged out successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Logout — revoke refresh token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/onboarding": {
      "patch": {
        "operationId": "AuthController_completeOnboarding",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Onboarding marked as completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark onboarding as completed",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/api/v1/categories": {
      "get": {
        "operationId": "CategoriesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all categories for the current user",
        "tags": [
          "Categories"
        ]
      },
      "post": {
        "operationId": "CategoriesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new category",
        "tags": [
          "Categories"
        ]
      }
    },
    "/api/v1/categories/from-recommendation": {
      "post": {
        "operationId": "CategoriesController_createFromRecommendation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryFromRecommendationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Unknown recommendation key"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a category from a recommendation",
        "tags": [
          "Categories"
        ]
      }
    },
    "/api/v1/categories/{id}": {
      "patch": {
        "operationId": "CategoriesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Category not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a category",
        "tags": [
          "Categories"
        ]
      },
      "delete": {
        "operationId": "CategoriesController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Category deleted"
          },
          "404": {
            "description": "Category not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a category",
        "tags": [
          "Categories"
        ]
      }
    },
...
    "/api/v1/analytics/mood-overview": {
      "get": {
        "operationId": "AnalyticsController_getMoodOverview",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Mood overview with distribution, category stats, 30-day trend, and weekday insights",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MoodOverviewResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get global mood analytics overview with weekday behavioral insights",
        "tags": [
          "Analytics"
        ]
      }
    },
...
  },
  "info": {
    "title": "LifeSpan API",
    "description": "Backend API for Life Timeline — build a visual timeline of your life",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "example": "SecurePass1",
            "minLength": 8,
            "description": "Min 8 chars, 1 uppercase, 1 lowercase, 1 number"
          },
          "timezone": {
            "type": "string",
            "example": "Europe/Berlin",
            "default": "UTC"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "AuthUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "timezone": {
            "type": "string"
          },
          "onboardingCompleted": {
            "type": "boolean"
          },
          "tier": {
            "type": "string",
            "enum": [
              "FREE",
              "PRO",
              "PREMIUM"
            ],
            "example": "FREE"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "timezone",
          "onboardingCompleted",
          "tier",
          "createdAt"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string"
          },
          "refresh_token": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/AuthUserDto"
          }
        },
        "required": [
          "access_token",
          "refresh_token",
          "user"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "example": "securePassword123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "GoogleExchangeDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "One-time OAuth exchange code"
          }
        },
        "required": [
          "code"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refresh_token": {
            "type": "string",
            "description": "Refresh token received during login/register"
          }
        },
        "required": [
          "refresh_token"
        ]
      },
...
      "CategoryResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "isSystem": {
            "type": "boolean"
          },
          "order": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "name",
          "color",
          "isSystem",
          "order"
        ]
      },
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Work"
          },
          "color": {
            "type": "string",
            "example": "#FF5733",
            "description": "Hex color code"
          },
          "order": {
            "type": "number",
            "example": 0,
            "description": "Display order"
          }
        },
        "required": [
          "name",
          "color"
        ]
      },
      "CreateCategoryFromRecommendationDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "example": "work",
            "description": "Recommendation key"
          },
          "name": {
            "type": "string",
            "example": "Work",
            "description": "Translated display name"
          }
        },
        "required": [
          "key",
          "name"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Work"
          },
          "color": {
            "type": "string",
            "example": "#FF5733"
          },
          "order": {
            "type": "number",
            "example": 0
          }
        }
      },
...
      "MoodDistributionItemDto": {
        "type": "object",
        "properties": {
          "moodId": {
            "type": "string"
          },
          "moodName": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "count": {
            "type": "number"
          },
          "percentage": {
            "type": "number"
          }
        },
        "required": [
          "moodId",
          "moodName",
          "color",
          "count",
          "percentage"
        ]
      },
...
      "CategoryMoodSummaryDto": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "averageMoodScore": {
            "type": "number"
          }
        },
        "required": [
          "categoryId",
          "name",
          "averageMoodScore"
        ]
      },
      "TrendPointDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string"
          },
          "score": {
            "type": "number"
          }
        },
        "required": [
          "date",
          "score"
        ]
      },
      "WeekdayInsightDto": {
        "type": "object",
        "properties": {
          "weekday": {
            "type": "number",
            "description": "Day of week (0=Monday, 6=Sunday)"
          },
          "averageScore": {
            "type": "number",
            "description": "Average mood score for this weekday (0-10)"
          },
          "sampleSize": {
            "type": "number",
            "description": "Number of data points for this weekday"
          }
        },
        "required": [
          "weekday",
          "averageScore",
          "sampleSize"
        ]
      },
      "ActivityInsightDto": {
        "type": "object",
        "properties": {
          "weekday": {
            "type": "number",
            "description": "Day of week (0=Monday, 6=Sunday)"
          },
          "averageActivityScore": {
            "type": "number",
            "description": "Average activity score for this weekday"
          },
          "sampleSize": {
            "type": "number",
            "description": "Number of data points for this weekday"
          }
        },
        "required": [
          "weekday",
          "averageActivityScore",
          "sampleSize"
        ]
      },
      "VolatilityInsightDto": {
        "type": "object",
        "properties": {
          "weekday": {
            "type": "number",
            "description": "Day of week (0=Monday, 6=Sunday)"
          },
          "standardDeviation": {
            "type": "number",
            "description": "Standard deviation of mood scores"
          },
          "sampleSize": {
            "type": "number",
            "description": "Number of data points for this weekday"
          }
        },
        "required": [
          "weekday",
          "standardDeviation",
          "sampleSize"
        ]
      },
      "RecoveryInsightDto": {
        "type": "object",
        "properties": {
          "weekday": {
            "type": "number",
            "description": "Day of week (0=Monday, 6=Sunday)"
          },
          "recoveryRate": {
            "type": "number",
            "description": "Rate of recovery events (0-1)"
          },
          "recoveryEvents": {
            "type": "number",
            "description": "Number of recovery events"
          },
          "totalOccurrences": {
            "type": "number",
            "description": "Total occurrences of this weekday"
          }
        },
        "required": [
          "weekday",
          "recoveryRate",
          "recoveryEvents",
          "totalOccurrences"
        ]
      },
      "BurnoutInsightDto": {
        "type": "object",
        "properties": {
          "detected": {
            "type": "boolean",
            "description": "Whether a burnout pattern was detected"
          },
          "type": {
            "type": "string",
            "description": "Type of pattern detected",
            "example": "work_stress_pattern"
          },
          "confidence": {
            "type": "number",
            "description": "Confidence level (0-1)"
          }
        },
        "required": [
          "detected"
        ]
      },
      "WeekdayInsightsDto": {
        "type": "object",
        "properties": {
          "bestMoodDay": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/WeekdayInsightDto"
              }
            ]
          },
          "worstMoodDay": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/WeekdayInsightDto"
              }
            ]
          },
          "mostActiveDay": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ActivityInsightDto"
              }
            ]
          },
          "leastActiveDay": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ActivityInsightDto"
              }
            ]
          },
          "mostUnstableDay": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/VolatilityInsightDto"
              }
            ]
          },
          "recoveryIndex": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/RecoveryInsightDto"
              }
            ]
          },
          "burnoutPattern": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/BurnoutInsightDto"
              }
            ]
          }
        }
      },
      "MoodOverviewResponseDto": {
        "type": "object",
        "properties": {
          "totalDaysWithMood": {
            "type": "number"
          },
          "averageMoodScore": {
            "type": "number"
          },
          "moodDistribution": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MoodDistributionItemDto"
            }
          },
          "bestCategory": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/CategoryMoodSummaryDto"
              }
            ]
          },
          "worstCategory": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/CategoryMoodSummaryDto"
              }
            ]
          },
          "trendLast30Days": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TrendPointDto"
            }
          },
          "weekdayInsights": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/WeekdayInsightsDto"
              }
            ]
          }
        },
        "required": [
          "totalDaysWithMood",
          "averageMoodScore",
          "moodDistribution",
          "trendLast30Days"
        ]
      },
...
    }
  }
}